chore: adds accuracyRunStatus to snapshot entries
The new field `accuracyRunStatus` is supposed to help guard against
cases where jest might fail in between, maybe due to LLM rate limit
errors or something else, and we then have a partially saved state of an
accuracy run. With the new field `accuracyRunStatus` we should be able
to safely look for last runs where `accuracyRunStatus` is done and have
complete state of accuracy snapshot.