chore: allow adding custom parameter scorers
In some cases, find, aggregate, count, explain, deleteMany, etc we need
to grade extra provided arguments depending on the prompt itself.
Sometimes additional parameters are fine and sometimes they are not. For
example: increasing the keys in filter might lead to a different result
hence if any such thing happens, we should grade the accuracy as 0 and
not 0.75. To suppor this use-case, this commit introduces the idea of
a custom scorer that could be plugged in to accuracy scorer to provided
more controlled accuracy grading.

Additionally this commit reverts the default behaviour of handling added
parameters. Earlier we were marking newly added parameters as
hallucinations and hence grading 0.75. But now, after figuring out that
most of our tools don't even expect extra parameters, we are flipping
the switch and instead will now grade 0 when additional parameters are
specified, unless there is a scorer provided to handle the custom
scoring logic.