chore: Gracefully fail when the server is down
The server might be closed because the connection made the
server die or because it's a unit test. In that case, do
not fail, just log the error and continue.

This is not a critical feature, in case that we couldn't
sync once, we will try again at some point or just let
it be.