chore: Add support for custom verbosity to the explain plan tool
While queryPlanner is a good default, there are scenarios where we
need more verbose outputs to understand why a query is not acting as
expected. For example, executionStats can be useful to see rw locks,
disk usage and flow throttling.

I'm not sure how good are models understanding the whole
executionPlan, and maybe we need to remove information that might
not be relevant for debugging to reduce the context usage, but this is
a decent first approach and we can improve later.