chore(ci): add dependency checks and reorganize tangled dependencies
This adds a dep check using `knip`. We use `depcheck` in Compass but this is deprecated and knip is recommend instead. This adds a basic configuration and then uses it to untangle some of our dependencies. This includes:

- Moving dependencies into dev dependencies, most notably `mongodb`! which we actually only use for types in the code itself at the moment.
- Moving certain types, consts into helper utils as they are only relevant there.

This seems like a useful feature and allows us to avoid mixing up our scripting or testing code with production.