chore: NodeDriverServiceProvider.connect blocks, breaking previous contract
NodeDriverServiceProvider.connect would return a Promise that was
lazily connected. However, now, blocks until there is a ping from the
server, forcing a connection.

This would usually be fine, however, in OIDC it would mean blocking
the whole MCP flow until the OIDC flow is connected. With the previous
version, this was already handled, but now we need to split the logic
a bit more to ensure that OIDC flows stay async.