feat(search): Add a new tool to list search and vector search indexes
It detects if the cluster supports search indexes the same way we
do in all other tooling: by catching the exception. In some old
clusters it might still return an empty list for unsupported
custers, but we can't really know for sure.

As we depend on mongot, we need to use the Atlas local
image, which is not supported by mongodb-runner. Instead,
we use testcontainers, that allows to run any docker image,
which is enough for our use case.

For now only run docker tests in Ubuntu
* macos-latest does not support nested virtualisation yet
* we don't have windows containers for atlas images