chore: change interface of describeWithMongoDB
The number of parameters were increasing as I wanted to add elicitInput
as another argument. This commit compacts the rest of the config
parameters in an object and modifies the usage to call the method
correctly.

Additionally, in places where we were just doing the defaults, this
commit modifies those calls to not pass anything and rely on default.