fix: fixes vector stage input parsing
Users were unable to use VectorSearch stage because Zod was using our
catch all stage (AnyAggregateStage) before VectorSearchStage to parse
and validate even the vector seach stage input and it passed through
because that schema is a catch all schema.

Because of that, in the input that we received, outputDimension was not
transformed ever.

This commit changes the order of application of schema so that
VectorSeachStage schema is validated first and then the catch all stage
schema.